pyproject.toml
[build-system]
requires = ["hatchling>=1.24", "hatch-vcs>=0.4"]
build-backend = "hatchling.build"

[project]
name = "mhz14a"
dynamic = ["version"]
description = "Python library for MH-Z14A CO₂ sensor"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    {name = "Osman Aslan", email = "admin@oaslananka.me"},
]
keywords = ["mhz14a", "co2", "sensor", "uart", "raspberry-pi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: System :: Hardware",
    "Typing :: Typed",
]
dependencies = [
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
    "build>=0.10.0",
]

[project.scripts]
mhz14a = "mhz14a.cli:main"

[project.urls]
Homepage = "https://github.com/oaslananka/mhz14a"
Repository = "https://github.com/oaslananka/mhz14a.git"
Issues = "https://github.com/oaslananka/mhz14a/issues"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["src/mhz14a"]

[tool.ruff]
line-length = 88
target-version = "py38"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["B011"]

[tool.mypy]
strict = true
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
